new/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
datasets>=2.0.0
Pillow>=8.0.0
numpy>=1.21.0
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0
nltk>=3.7
